moved some classes and added ManageStoreNavigator
