+++ v8engine/src/main/java/cm/aptoide/pt/v8engine/view/account/ImagePickerPresenter.java
@@ -6,8 +6,7 @@ import android.support.annotation.NonNull;
 import cm.aptoide.pt.v8engine.crashreports.CrashReport;
 import cm.aptoide.pt.v8engine.presenter.Presenter;
 import cm.aptoide.pt.v8engine.presenter.View;
-import cm.aptoide.pt.v8engine.view.account.store.ImagePickerNavigator;
-import cm.aptoide.pt.v8engine.view.account.store.exception.InvalidImageException;
+import cm.aptoide.pt.v8engine.view.account.exception.InvalidImageException;
 import cm.aptoide.pt.v8engine.view.permission.AccountPermissionProvider;
 import rx.Completable;
 import rx.Observable;
