+++ v8engine/src/main/java/cm/aptoide/pt/v8engine/view/account/exception/StoreCreationException.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.v8engine.view.account.store.exception;
+package cm.aptoide.pt.v8engine.view.account.exception;
 
 public class StoreCreationException extends Exception {
