+++ v8engine/src/main/java/cm/aptoide/pt/v8engine/view/account/store/StoreManager.java
@@ -13,8 +13,8 @@ import cm.aptoide.pt.dataprovider.ws.v3.CheckUserCredentialsRequest;
 import cm.aptoide.pt.dataprovider.ws.v7.SetStoreImageRequest;
 import cm.aptoide.pt.dataprovider.ws.v7.SimpleSetStoreRequest;
 import cm.aptoide.pt.dataprovider.ws.v7.store.RequestBodyFactory;
-import cm.aptoide.pt.v8engine.view.account.store.exception.InvalidImageException;
-import cm.aptoide.pt.v8engine.view.account.store.exception.StoreCreationException;
+import cm.aptoide.pt.v8engine.view.account.exception.InvalidImageException;
+import cm.aptoide.pt.v8engine.view.account.exception.StoreCreationException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Collections;
 import java.util.List;
