+++ v8engine/src/main/java/cm/aptoide/pt/v8engine/view/account/user/ManageUserFragment.java
@@ -35,8 +35,8 @@ import cm.aptoide.pt.v8engine.view.account.ImagePickerPresenter;
 import cm.aptoide.pt.v8engine.view.account.ImageValidator;
 import cm.aptoide.pt.v8engine.view.account.PhotoFileGenerator;
 import cm.aptoide.pt.v8engine.view.account.UriToPathResolver;
-import cm.aptoide.pt.v8engine.view.account.store.ImagePickerNavigator;
-import cm.aptoide.pt.v8engine.view.account.store.exception.InvalidImageException;
+import cm.aptoide.pt.v8engine.view.account.ImagePickerNavigator;
+import cm.aptoide.pt.v8engine.view.account.exception.InvalidImageException;
 import cm.aptoide.pt.v8engine.view.dialog.ImagePickerDialog;
 import cm.aptoide.pt.v8engine.view.permission.AccountPermissionProvider;
 import cm.aptoide.pt.v8engine.view.permission.PermissionProvider;
